renamed app.py to window.py
