+++ openpype/tools/creator/__init__.py
@@ -1,4 +1,4 @@
-from .app import (
+from .window import (
     show,
 )
